requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
passlib[bcrypt]
bcrypt==4.0.1
alembic
psycopg2-binary
tenacity
structlog
python-jose[cryptography]
pydantic-settings
